Adds variable number of panes to first window based on number of users
Adds a panel for each user with their associated icon and name. Since we
dont have a file yet with preset users it wont display anything.

The 0 on the side by the button shows number of users. This will be
removed once I can test with a valid settings file.